remove circular import of citation function
