+++ src/hf_hydrodata/point.py
@@ -1127,75 +1127,6 @@ def _construct_string_from_qparams(data_type, options):
     return result_string
 
 
-def _get_point_citations(dataset):
-    """
-    Return a dictionary with relevant citation information.
-
-    Parameters
-    ----------
-    dataset : str
-        Source from which requested data originated. Currently supported: 'usgs_nwis', 'snotel',
-        'scan', 'ameriflux', 'jasechko_2024', 'fan_2013'.
-
-    Returns
-    -------
-    str
-        String containing overall attribution instructions for the provided dataset.
-    """
-    try:
-        assert dataset in [
-            "usgs_nwis",
-            "snotel",
-            "scan",
-            "ameriflux",
-            "jasechko_2024",
-            "fan_2013",
-        ]
-    except:
-        raise ValueError(
-            f"Unexpected value of dataset, {dataset}. Supported values include 'usgs_nwis', 'snotel', 'scan', 'ameriflux', 'jasechko_2024', and 'fan_2013'"
-        )
-
-    if dataset == "usgs_nwis":
-        c = (
-            "Most U.S. Geological Survey (USGS) information resides in Public Domain and "
-            "may be used without restriction, though they do ask that proper credit be given. "
-            'An example credit statement would be: "(Product or data name) courtesy of the U.S. Geological Survey".\n'
-            "Source: https://www.usgs.gov/information-policies-and-instructions/acknowledging-or-crediting-usgs"
-        )
-
-    elif dataset in ["snotel", "scan"]:
-        c = (
-            "Most information presented on the USDA Web site is considered public domain information. "
-            "Public domain information may be freely distributed or copied, but use of appropriate "
-            "byline/photo/image credits is requested. Attribution may be cited as follows: "
-            '"U.S. Department of Agriculture"\nSource: https://www.usda.gov/policies-and-links'
-        )
-
-    elif dataset == "ameriflux":
-        c = (
-            "All AmeriFlux sites provided by the HydroData service follow the CC-BY-4.0 License. "
-            "The CC-BY-4.0 license specifies that the data user is free to Share (copy and "
-            "redistribute the material in any medium or format) and/or Adapt (remix, transform, "
-            "and build upon the material) for any purpose. "
-            "Users of this data must acknowledge the AmeriFlux data resource with the "
-            'following statement: "Funding for the AmeriFlux data portal was provided by the U.S. '
-            'Department of Energy Office of Science." '
-            "Additionally, for each AmeriFlux site used, you must provide a citation to the site "
-            "data product that includes the data product DOI. The DOI for each site is included in the "
-            "DataFrame returned by the hf_hydrodata get_point_metadata method, in the doi column.\n"
-            "Source: https://ameriflux.lbl.gov/data/data-policy/"
-        )
-
-    elif dataset == "jasechko_2024":
-        c = "Dataset DOI: 10.1038/s41586-023-06879-8"
-
-    elif dataset == "fan_2013":
-        c = "Dataset DOI: 10.1126/science.1229881"
-
-    return c
-
-
 def _validate_user():
     email, pin = _get_registered_api_pin()
     url_security = f"{HYDRODATA_URL}/api/api_pins?pin={pin}&email={email}"
